purchase_data_logsの追加に伴い、編集ページのデザイン変更

+++ backend/src/handler/addItem.go
@@ -145,7 +145,7 @@ func AddItem(c echo.Context) error {
 		UserItemID: maxItemID.UserItemID + 1,
 		LatestReminder: false,
 		UntilNextTimeByDays: untilNextTimeByDays,
-		AveragePrice: averagePrice,
+		AveragePrice: float32(averagePrice),
 	}
 
 	purchaseDateLogs := &schema.PurchaseDataLogs{
